Wrote ColorTest stuff to try implementing a pixel-based display. Also expanded the volume, and the fact that it manually assigns pixel colors likely slows things down a lot. I probably should make it write to an array that gets dumped into the image each round. Later.
Wrote it so that it loads up an array with rgb data first, then sends that into the image. I assume this works faster than individually setting each pixel as its corresponding face gets handled. It turns 5 random switches on every pass through the loop, and after 1000 passes it waits for 500 rounds before repeating the process.
There is a glitch somewhere that causes a repeating activation chain event under the right circumstances. I haven't had it crash the system yet, but it might get triggered if this runs long enough.

Note: As a 22z, 80y, 30x, with 125 faces on each one, there are 59,400,000 switches. I thought that was kinda interesting. Didn't expect to cross 1,000,000 any time soon really.